container of notification made
